Moved plotting code for transit_fitting (although it appears unused).
