+++ stellarphot/gui_tools/comparison_functions.py
@@ -19,9 +19,9 @@ except ImportError:
 
 from stellarphot.differential_photometry import *
 from stellarphot.photometry import *
-from stellarphot.io import TessSubmission, TOI, TessTargetFile
 from stellarphot.gui_tools.seeing_profile_functions import set_keybindings
 from stellarphot.gui_tools.fits_opener import FitsOpener
+from stellarphot.io import TessSubmission, TOI, TessTargetFile
 from stellarphot.utils.comparison_utils import read_file, set_up, crossmatch_APASS2VSX, mag_scale, in_field
