+++ stellarphot/io/tests/test_aij_io.py
@@ -1,5 +1,3 @@
-from pathlib import Path
-
 import numpy as np
 
 from astropy.coordinates import SkyCoord
