+++ stellarphot/transit_fitting/gui/tests/test_transit_fitting_gui.py
