+++ stellarphot/transit_fitting/plotting/__init__.py
@@ -1,4 +1,3 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
 
-from .aij import *
-from .tess import *
+from .ploting_core import *
